refactoring: Move user filter logic to separate file for modularity
