+++ internal/delivery/api/endpoint.go
@@ -161,14 +161,14 @@ const (
 	GetPolicyNotification
 
 	// Stack
-	GetStacks           // 스택관리/조회
 	CreateStack         // 스택관리/생성
-	CheckStackName      // 스택관리/조회
+	GetStacks           // 스택관리/조회
 	GetStack            // 스택관리/조회
 	UpdateStack         // 스택관리/수정
 	DeleteStack         // 스택관리/삭제
-	GetStackKubeConfig  // 스택관리/조회
+	CheckStackName      // 스택관리/조회
 	GetStackStatus      // 스택관리/조회
+	GetStackKubeConfig  // 스택관리/조회
 	SetFavoriteStack    // 스택관리/조회
 	DeleteFavoriteStack // 스택관리/조회
 	InstallStack        // 스택관리 / 조회
