+++ internal/middleware/auth/authorizer/authorizer.go
@@ -22,6 +22,7 @@ func NewDefaultAuthorization(repo repository.Repository) *defaultAuthorization {
 	//d.addFilters(RBACFilter)
 	//d.addFilters(RBACFilterWithEndpoint)
 	d.addFilters(AdminApiFilter)
+	d.addFilters(OrganizationFilter)
 
 	return d
 }
